requirements.txt
# VerbatimAI - Enterprise Requirements
# Core Framework & Dependencies - STABLE VERSIONS
numpy==1.24.3
pandas==2.1.3
streamlit==1.28.1
plotly==5.17.0
assemblyai==0.21.0
requests==2.31.0
python-dotenv==1.0.0

# Advanced Analytics & ML
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
spacy>=3.4.0

# Document Processing & Export
python-docx==1.1.0
reportlab==4.0.7
pydub==0.25.1
jinja2==3.1.2
weasyprint==60.2

# Audio/Video Processing
pyaudio==0.2.11
opencv-python==4.8.1.78
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0

# Background Processing & Async
redis==5.0.1
celery==5.3.4
flower==2.0.1
kombu==5.3.4

# Enhanced Visualization & UI
seaborn==0.13.0
altair==5.2.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-elements==0.1.0
streamlit-lottie==0.0.5

# Email & Communication
smtplib2==0.2.1
email-validator==2.1.0
sendgrid==6.10.0

# Database & Storage
sqlalchemy==2.0.23
sqlite3
psycopg2-binary==2.9.9

# Security & Authentication
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.1.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Performance & Optimization
cachetools==5.3.2
memory-profiler==0.61.0
psutil==5.9.6

# Additional Utilities
python-dateutil==2.8.2
pytz==2023.3
jsonschema==4.20.0
pyyaml==6.0.1
tqdm==4.66.1

# API & Web Framework (for future API endpoints)
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2

# Cloud Storage (optional)
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0 